Playlists : Mandatory Save() to persist data; support for writing relative paths

+++ ATL.unit-test/Playlist/ASXIO.cs
@@ -1,60 +1,21 @@
 ﻿using ATL.Playlist;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Xml;
 
 namespace ATL.test.IO.Playlist
 {
     [TestClass]
-    public class ASXIO
+    public class ASXIO : PlaylistIOTest
     {
         [TestMethod]
         public void PLIO_R_ASX()
         {
-            var testFileLocation = TestUtils.CopyFileAndReplace(TestUtils.GetResourceLocationRoot() + "_Playlists/playlist.asx", "$PATH", TestUtils.GetResourceLocationRoot(false).Replace("\\", "/"));
-
-            try
-            {
-                var theReader = PlaylistIOFactory.GetInstance().GetPlaylistIO(testFileLocation);
-                var filePaths = theReader.FilePaths;
-                bool foundHttp = false;
-
-                Assert.IsNotInstanceOfType(theReader, typeof(ATL.Playlist.IO.DummyIO));
-                Assert.AreEqual(5, filePaths.Count);
-                foreach (string s in theReader.FilePaths)
-                {
-                    if (!s.StartsWith("http", StringComparison.InvariantCultureIgnoreCase)) Assert.IsTrue(System.IO.File.Exists(s));
-                    else foundHttp = true;
-                }
-                Assert.IsTrue(foundHttp);
-                foreach (Track t in theReader.Tracks)
-                {
-                    // Ensure the track has been parsed when it points to a file
-                    if (!t.Path.StartsWith("http", StringComparison.InvariantCultureIgnoreCase)) Assert.IsTrue(t.Duration > 0);
-                }
-            }
-            finally
-            {
-                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
-            }
+            PLIO_R("playlist.asx", TestUtils.GetResourceLocationRoot(false).Replace("\\", "/"), 5);
         }
 
         [TestMethod]
         public void PLIO_W_ASX()
         {
-            IList<string> pathsToWrite = new List<string>();
-            string testTrackLocation1 = TestUtils.CopyAsTempTestFile("MP3/empty.mp3");
-            string testTrackLocation2 = TestUtils.CopyAsTempTestFile("MOD/mod.mod");
-            pathsToWrite.Add(testTrackLocation1);
-            pathsToWrite.Add(testTrackLocation2);
-            pathsToWrite.Add("http://this-is-a-stre.am:8405/live");
-
-            IList<Track> tracksToWrite = new List<Track>();
-            foreach (var s in pathsToWrite) tracksToWrite.Add(new Track(s));
-
-
             string testFileLocation = TestUtils.CreateTempTestFile("test.asx");
             bool defaultPathSetting = ATL.Settings.PlaylistWriteAbsolutePath;
             try
@@ -64,6 +25,8 @@ namespace ATL.test.IO.Playlist
                 // Test Path writing + absolute formatting
                 ATL.Settings.PlaylistWriteAbsolutePath = true;
                 pls.FilePaths = pathsToWrite;
+                pls.Save();
+
                 IList<string> parents = new List<string>();
                 int index = -1;
 
@@ -101,6 +64,8 @@ namespace ATL.test.IO.Playlist
                 // Test Track writing + relative formatting
                 ATL.Settings.PlaylistWriteAbsolutePath = false;
                 pls.Tracks = tracksToWrite;
+                pls.Save();
+
                 index = -1;
                 parents.Clear();
 
@@ -136,12 +101,61 @@ namespace ATL.test.IO.Playlist
             finally
             {
                 ATL.Settings.PlaylistWriteAbsolutePath = defaultPathSetting;
-                if (Settings.DeleteAfterSuccess)
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
+            }
+        }
+
+        [TestMethod]
+        public void PLIO_PLIO_RW_Absolute_Relative_Path_ASX()
+        {
+            var testFileLocation = PLIO_RW_Absolute_Relative_Path("asx");
+            try
+            {
+                IList<string> parents = new List<string>();
+                int nbEntries = -1;
+
+                using FileStream fs = new FileStream(testFileLocation, FileMode.Open);
+                using XmlReader source = XmlReader.Create(fs);
+                while (source.Read())
                 {
-                    File.Delete(testTrackLocation1);
-                    File.Delete(testTrackLocation2);
-                    File.Delete(testFileLocation);
+                    if (source.NodeType != XmlNodeType.Element) continue;
+
+                    if (source.Name.Equals("asx", StringComparison.OrdinalIgnoreCase)) parents.Add(source.Name.ToLower());
+                    else if (source.Name.Equals("entry", StringComparison.OrdinalIgnoreCase) && parents.Contains("asx"))
+                    {
+                        nbEntries++;
+                        parents.Add(source.Name.ToLower());
+                    }
+                    else if (parents.Contains("entry"))
+                    {
+                        if (source.Name.Equals("ref", StringComparison.OrdinalIgnoreCase))
+                        {
+                            var writtenPath = source.GetAttribute("HREF");
+                            switch (nbEntries)
+                            {
+                                case 0:
+                                    Assert.AreEqual(remoteFilePath, writtenPath);
+                                    break;
+                                case 1:
+                                    Assert.AreEqual(localFilePath1, writtenPath);
+                                    break;
+                                case 2:
+                                    Assert.AreEqual(TestUtils.MakePathRelative(testFileLocation, localFilePath2),
+                                        writtenPath);
+                                    break;
+                            }
+                        }
+                        else if (source.Name.Equals("title", StringComparison.OrdinalIgnoreCase) && 1 == nbEntries)
+                        {
+                            Assert.AreEqual(NEW_TITLE, getXmlValue(source));
+                        }
+                    }
                 }
+                Assert.AreEqual(3, nbEntries + 1);
+            }
+            finally
+            {
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
             }
         }
