+++ ATL.unit-test/Playlist/PLSIO.cs
@@ -1,55 +1,19 @@
 ﻿using ATL.Playlist;
-using System;
 
 namespace ATL.test.IO.Playlist
 {
     [TestClass]
-    public class PLSIO
+    public class PLSIO : PlaylistIOTest
     {
         [TestMethod]
         public void PLIO_R_PLS()
         {
-            string testFileLocation = TestUtils.CopyFileAndReplace(TestUtils.GetResourceLocationRoot() + "_Playlists/playlist.pls", "$PATH", TestUtils.GetResourceLocationRoot(false));
-
-            try
-            {
-                IPlaylistIO theReader = PlaylistIOFactory.GetInstance().GetPlaylistIO(testFileLocation);
-                bool foundHttp = false;
-
-                Assert.IsNotInstanceOfType(theReader, typeof(ATL.Playlist.IO.DummyIO));
-                Assert.AreEqual(5, theReader.FilePaths.Count);
-                foreach (string s in theReader.FilePaths)
-                {
-                    if (!s.StartsWith("http", StringComparison.InvariantCultureIgnoreCase)) Assert.IsTrue(System.IO.File.Exists(s));
-                    else foundHttp = true;
-                }
-                Assert.IsTrue(foundHttp);
-                foreach (Track t in theReader.Tracks)
-                {
-                    // Ensure the track has been parsed when it points to a file
-                    if (!t.Path.StartsWith("http", StringComparison.InvariantCultureIgnoreCase)) Assert.IsTrue(t.Duration > 0);
-                }
-            }
-            finally
-            {
-                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
-            }
+            PLIO_R("playlist.pls", TestUtils.GetResourceLocationRoot(false), 5);
         }
 
         [TestMethod]
         public void PLIO_W_PLS()
         {
-            IList<string> pathsToWrite = new List<string>();
-            string testTrackLocation1 = TestUtils.CopyAsTempTestFile("MP3/empty.mp3");
-            string testTrackLocation2 = TestUtils.CopyAsTempTestFile("MOD/mod.mod");
-            pathsToWrite.Add(testTrackLocation1);
-            pathsToWrite.Add(testTrackLocation2);
-            pathsToWrite.Add("http://this-is-a-stre.am:8405/live");
-
-            IList<Track> tracksToWrite = new List<Track>();
-            foreach (var s in pathsToWrite) tracksToWrite.Add(new Track(s));
-
-
             string testFileLocation = TestUtils.CreateTempTestFile("test.pls");
             bool defaultPathSetting = ATL.Settings.PlaylistWriteAbsolutePath;
             try
@@ -59,6 +23,7 @@ namespace ATL.test.IO.Playlist
                 // Test Path writing + absolute formatting
                 ATL.Settings.PlaylistWriteAbsolutePath = true;
                 pls.FilePaths = pathsToWrite;
+                pls.Save();
 
                 using (FileStream fs = new FileStream(testFileLocation, FileMode.Open))
                 {
@@ -76,7 +41,7 @@ namespace ATL.test.IO.Playlist
                         int counter = 1;
                         foreach (string s in pathsToWrite)
                         {
-                            Assert.AreEqual("File" + counter + "="+ s, sr.ReadLine());
+                            Assert.AreEqual("File" + counter + "=" + s, sr.ReadLine());
                             Assert.AreEqual("Title" + counter + "=" + System.IO.Path.GetFileNameWithoutExtension(s), sr.ReadLine());
                             Assert.AreEqual("Length" + counter + "=-1", sr.ReadLine());
                             sr.ReadLine();
@@ -96,6 +61,7 @@ namespace ATL.test.IO.Playlist
                 // Test Track writing + relative formatting
                 ATL.Settings.PlaylistWriteAbsolutePath = false;
                 pls.Tracks = tracksToWrite;
+                pls.Save();
 
                 using (FileStream fs = new FileStream(testFileLocation, FileMode.Open))
                 using (StreamReader sr = new StreamReader(fs))
@@ -125,12 +91,52 @@ namespace ATL.test.IO.Playlist
             finally
             {
                 ATL.Settings.PlaylistWriteAbsolutePath = defaultPathSetting;
-                if (Settings.DeleteAfterSuccess)
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
+            }
+        }
+
+        [TestMethod]
+        public void PLIO_PLIO_RW_Absolute_Relative_Path_PLS()
+        {
+            var testFileLocation = PLIO_RW_Absolute_Relative_Path("pls");
+            try
+            {
+                int nbEntries = 1;
+                using (FileStream fs = new FileStream(testFileLocation, FileMode.Open))
+                using (StreamReader sr = new StreamReader(fs))
                 {
-                    File.Delete(testTrackLocation1);
-                    File.Delete(testTrackLocation2);
-                    File.Delete(testFileLocation);
+                    sr.ReadLine();
+                    for (int i = 0; i < 3; i++)
+                    {
+                        string path;
+                        switch (nbEntries)
+                        {
+                            case 1:
+                                path = remoteFilePath;
+                                break;
+                            case 2:
+                                path = localFilePath1;
+                                break;
+                            case 3:
+                                path = TestUtils.MakePathRelative(testFileLocation, localFilePath2);
+                                break;
+                            default:
+                                path = "";
+                                break;
+                        }
+                        sr.ReadLine();
+                        Assert.AreEqual("File" + nbEntries + "=" + path, sr.ReadLine());
+                        var title = sr.ReadLine();
+                        if (2 == nbEntries) Assert.AreEqual("Title" + nbEntries + "=" + NEW_TITLE, title);
+                        sr.ReadLine(); // Duration
+                        nbEntries++;
+                    }
                 }
+                Assert.AreEqual(4, nbEntries);
+            }
+            finally
+            {
+                if (Settings.DeleteAfterSuccess) File.Delete(testFileLocation);
             }
         }
     }
